Behind the programs: see extended desctiption
The program keeps the doors open for a short period of time, then the doors close. When that happens, if more people come (which is what the volume determines. Its how quickly they come), they will just wait in line. I didn't show that on the output since I figured it would have too much going on with it. It only iterates a few times before terminating since I didn't want to show this forever. It will essentially stop having people generated and get rid of everyone already in line then stop running.